+++ authority-portal-frontend/src/app/pages/component-registration-pages/register-connector-page/register-connector-page/register-connector-page.component.ts
@@ -33,6 +33,7 @@ export class RegisterConnectorPageComponent implements OnInit, OnDestroy {
 
   createActionName = 'Register Connector';
   backLink = '/my-organization/connectors/new';
+  exitLink = '/my-organization/connectors';
 
   @ViewChild('stepper') stepper!: MatStepper;